separate admins pages from public pages

+++ src/controllers/Item.h
@@ -14,9 +14,10 @@ namespace api::v1 {
         METHOD_ADD(Item::getList, "", drogon::Get, drogon::Options);
         METHOD_ADD(Item::getOne, "{1}", drogon::Get, drogon::Options);
         METHOD_ADD(Item::getListAdmin, "admin", drogon::Get, drogon::Options, "api::v1::filters::JwtFilter");
-        METHOD_ADD(Item::createItem, "", drogon::Post, drogon::Options, "api::v1::filters::JwtFilter");
-        METHOD_ADD(Item::updateItem, "{1}", drogon::Put, drogon::Options, "api::v1::filters::JwtFilter");
-        METHOD_ADD(Item::deleteItem, "{1}", drogon::Delete, drogon::Options, "api::v1::filters::JwtFilter");
+        METHOD_ADD(Item::getOne, "admin/{1}", drogon::Get, drogon::Options, "api::v1::filters::JwtFilter");
+        METHOD_ADD(Item::createItem, "admin", drogon::Post, drogon::Options, "api::v1::filters::JwtFilter");
+        METHOD_ADD(Item::updateItem, "admin/{1}", drogon::Put, drogon::Options, "api::v1::filters::JwtFilter");
+        METHOD_ADD(Item::deleteItem, "admin/{1}", drogon::Delete, drogon::Options, "api::v1::filters::JwtFilter");
         METHOD_LIST_END
         [[nodiscard]] static Json::Value getJsonResponse(const drogon::orm::Result &r);
         void getListAdmin(
